Module tox::toxcore::crypto_core

Functions for the core crypto.

Reexports

pub use sodiumoxide::crypto::box_::*;

Functions

crypto_init
Run before using crypto.

decrypt_data_symmetric
Returns plain data from encrypted, with length of encrypted - 16 due to padding, or () if data couldn't be decrypted.

encrypt_data_symmetric
Returns encrypted data from plain, with length of plain + 16 due to padding.

encrypt_precompute
Precomputes the shared key from their_public_key and our_secret_key.

increment_nonce
Inrement given nonce by 1.

increment_nonce_number
Inrement given nonce by number num.

public_key_valid
Check if Tox public key PUBLICKEYBYTES is valid. Should be used only for input validation.

random_u32
Return a random number.

random_u64
Return a random number.

randombytes_into
randombytes_into() fills a buffer buf with random data.