Module tox::toxcore::dht

DHT part of the toxcore.

takes care of the serializing and de-serializing DHT packets

..

Structs

Bucket
Structure for holding nodes.

BucketIter
Iterator over Bucket.

DhtPacket
Standard DHT packet that encapsulates in the encrypted payload DhtPacketT.

DhtRequest
DHT Request packet structure.

GetNodes
Request to get address of given DHT PK, or nodes that are closest in DHT to the given PK.

Kbucket
K-buckets structure to hold up to KBUCKET_MAX_ENTRIES * BUCKET_DEFAULT_SIZE nodes close to own PK.

KbucketIter
Iterator over PackedNodes in Kbucket.

NatPingReq
NAT Ping; used to see if a friend we are not connected to directly is online and ready to do the hole punching.

NatPingResp
NAT Ping; used to see if a friend we are not connected to directly is online and ready to do the hole punching.

PackedNode
PackedNode format is a way to store the node info in a small yet easy to parse format.

PingReq
Used to request/respond to ping. Use in an encrypted form.

PingResp
Used to request/respond to ping. Use in an encrypted form.

SendNodes
Response to GetNodes request, containing up to 4 nodes closest to the requested node.

Enums

IpType
Used by PackedNode.

Constants

BUCKET_DEFAULT_SIZE
Default number of nodes that bucket can hold.

DHT_PACKET_MIN_SIZE
Minimal size of DhtPacket in bytes.

GET_NODES_SIZE
Size of serialized GetNodes in bytes.

KBUCKET_BUCKETS
Default number of Buckets that Kbucket holds.

KBUCKET_MAX_ENTRIES
Maximum number of Buckets that Kbucket can hold.

NAT_PING_SIZE
Length in bytes of NatPings when serialized into bytes.

NAT_PING_TYPE
NatPing type byte for NatPingReq and NatPingResp.

PACKED_NODE_IPV4_SIZE
Size in bytes of serialized PackedNode with IPv4.

PACKED_NODE_IPV6_SIZE
Size in bytes of serialized PackedNode with IPv6.

PING_SIZE
Length in bytes of PingReq and PingResp when serialized into bytes.

Traits

DhtPacketT
Trait for types of DHT packets that can be put in DhtPacket.

DhtRequestT
Trait for types of DHT requests that can be put in DhtRequest.

Distance
Trait for functionality related to distance between PublicKeys.

Functions

kbucket_index
Calculate the k-bucket index of a PK compared to "own" PK.