Struct tox::toxcore::dht::DhtPacket

pub struct DhtPacket {
    pub sender_pk: PublicKey,
    // some fields omitted
}

Standard DHT packet that encapsulates in the encrypted payload DhtPacketT.

Length 	Contents

1 	uint8_t PacketKind

32 	Sender DHT Public Key

24 	Random nonce

variable 	Encrypted payload

PacketKind values for DhtPacket can be only <= 4.

https://zetok.github.io/tox-spec/#dht-packet

Fields
sender_pk: PublicKey

Public key of sender.

Methods

impl DhtPacket

pub fn new<P>(
symmetric_key: &PrecomputedKey,
own_public_key: &PublicKey,
nonce: &Nonce,
dp: &P
) -> Self where
P: DhtPacketT,

Create new DhtPacket with bytes.

pub fn kind(&self) -> PacketKind

Get PacketKind that DhtPacket's payload is supposed to contain.

pub fn get_payload<P>(&self, own_secret_key: &SecretKey) -> Option<P> where
P: DhtPacketT,

Decrypt payload and try to parse it as packet type.

To get info about it's packet type use .kind() method.

Returns None in case of faliure:

fails to decrypt

fails to parse as given packet type

pub fn ping_resp(
&self,
secret_key: &SecretKey,
symmetric_key: &PrecomputedKey,
own_public_key: &PublicKey
) -> Option<Self>

Create DHT Packet with Ping response to Ping request that packet contained.

Nonce for the response is automatically generated.