Struct tox::toxcore::dht::Kbucket

pub struct Kbucket { /* fields omitted */ }

K-buckets structure to hold up to KBUCKET_MAX_ENTRIES * BUCKET_DEFAULT_SIZE nodes close to own PK.

Nodes in bucket are sorted by closeness to the PK; closest node is the first, while furthest is last.

Further reading: Tox spec.

Methods

impl Kbucket

pub fn new(n: u8, pk: &PublicKey) -> Self

Create a new Kbucket.

n – number of Buckets held.

pub fn size(&self) -> u8

Number of Buckets held.

pub fn try_add(&mut self, node: &PackedNode) -> bool

Add PackedNode to Kbucket.

Node can be added only if:

its kbucket index is lower than the number of buckets.

Bucket to which it is added has free space or added node is closer to the PK than other node in the bucket.

Returns true if node was added successfully, false otherwise.

pub fn remove(&mut self, node_pk: &PublicKey)

Remove PackedNode with given PK from the Kbucket.

pub fn get_closest(&self, pk: &PublicKey) -> Vec<PackedNode>

Get (up to) 4 closest nodes to given PK.

Functionality for SendNodes.

Returns less than 4 nodes only if Kbucket contains less than 4 nodes.

pub fn contains(&self, pk: &PublicKey) -> bool

Check if Kbucket contains PackedNode with given PK.

pub fn can_add(&self, pk: &PublicKey) -> bool

Naive check whether a PackedNode can be added to the Kbucket.

Returns true if Bucket where node could be placed is not full and node is not already in the Bucket.

Otherwise false is returned.

pub fn is_empty(&self) -> bool

Check if Kbucket is empty.

Returns true if all buckets are empty, false otherwise.

Important traits for KbucketIter<'a>
impl<'a> Iterator for KbucketIter<'a> type Item = &'a PackedNode;

pub fn iter(&self) -> KbucketIter

Create iterator over PackedNodes in Kbucket.