Struct tox::toxcore::dht::PackedNode

pub struct PackedNode { /* fields omitted */ }

PackedNode format is a way to store the node info in a small yet easy to parse format.

It is used in many places in Tox, e.g. in DHT Send nodes.

To store more than one node, simply append another on to the previous one:

[packed node 1][packed node 2][...]

Serialized Packed node:

Length 	Content

1 	IpType

4 or 16 	IPv4 or IPv6 address

2 	port

32 	node ID

Size of serialized PackedNode is 39 bytes with IPv4 node info, or 51 with IPv6 node info.

DHT module should use only UDP variants of IpType, given that DHT runs solely on the UDP.

Methods

impl PackedNode

pub fn new(udp: bool, saddr: SocketAddr, pk: &PublicKey) -> Self

New PackedNode.

udp - whether UDP or TCP should be used. UDP is used for DHT nodes, whereas TCP is used for TCP relays. When true, UDP is used, otherwise TCP is used.

pub fn ip_type(&self) -> IpType

Get an IP type from the PackedNode.

pub fn ip(&self) -> IpAddr

Get an IP address from the PackedNode.

pub fn socket_addr(&self) -> SocketAddr

Get a Socket address from the PackedNode.

pub fn pk(&self) -> &PublicKey

Get an IP address from the PackedNode.