Struct tox::toxcore::dht_new::packet::PingResponse

pub struct PingResponse {
    pub id: u64,
}

Used to request/respond to ping. Used in an encrypted form.

Used in:

DhtPacket

DhtRequest

Serialized form:

Ping Packet (request and response)

Packet type 0x01 for response.

Response ID must match ID of the request, otherwise ping is invalid.

Length 	Contents

1 	0x01

8 	Ping ID

Serialized form should be put in the encrypted part of DHT packet.

PingResponse can only be created as a response to PingRequest.

Fields
id: u64

Ping id same as requested from PingRequest