Module tox::toxcore::dht_node

Functionality needed to work as a DHT node.

Made on top of dht and network modules.

Macros

task_local
A macro to create a static of type LocalKey

try_ready
A macro for extracting the successful type of a Poll<T, E>.

Structs

DhtNode
Own DHT node data.

ToxCodec
Struct to use for {de-,}serializing Tox UDP packets.

Functions

receive_packets
Spawn a thread that will start receiving packets from ToxSplitStream.

send_packets
Spawn a thread that will start sending packets via ToxSplitSink.

Type Definitions

SendSink
Type representing future Send via SplitSink.

ToxRecvUdpPacket
Type representing received Tox UDP packets.

ToxSplitSink
Type for sending SplitSink with ToxCodec.

ToxSplitStream
Type for receiving SplitStream with ToxCodec.

ToxUdpPacket
Type representing Tox UDP packets.