Struct tox::toxcore::dht_node::DhtNode

pub struct DhtNode { /* fields omitted */ }

Own DHT node data.

Contains:

DHT public key

DHT secret key

Close List (Kbucket with nodes close to own DHT public key)

ping timeout lists (TimeoutQueue)

Adding node to Close List

Before a PackedNode is added to the Close List, it needs to be checked whether:

it can be added to Kbucket (using Kbucket::can_add())

PackedNode is actually online

Once the first check passes node is added to the temporary list, and a GetNodes request is sent to it in order to check whether it's online. If the node responds correctly within PING_TIMEOUT, it's removed from temporary list and added to the Close List.

Methods

impl DhtNode

pub fn new() -> Result<Self>

Create new DhtNode instance.

Note: a new instance generates new DHT public and secret keys.

DHT PublicKey and SecretKey are supposed to be ephemeral.

pub fn try_add(&mut self, node: &PackedNode) -> bool

Try to add nodes to Kbucket.

Wrapper around Kbucket's method.

pub fn remove_timed_out(&mut self, secs: u64)

Remove nodes that have crossed secs timeout threshold.

pub fn request_ping(&self, peer: &PackedNode) -> ToxUdpPacket

Create a ToxUdpPacket with request for ping response from a peer.

ToxUdpPacket is to be passed to Sender created by send_packets().

pub fn respond_ping(
&self,
sink: ToxSplitSink,
peer_addr: SocketAddr,
request: &DhtPacket
) -> Option<SendSink>

Create a future sending DhtPacket that encapsulates ping response to supplied ping request.

pub fn create_getn(&self, peer_pk: &PublicKey) -> (RequestId, DhtPacket)

Create a DhtPacket to peer's PublicKey containing a GetNodes request for nodes close to own DHT PublicKey.

RequestId is to be used for tracking node timeouts.

pub fn request_nodes(&mut self, peer: &PackedNode) -> (RequestId, ToxUdpPacket)

Create a ToxUdpPacket with request for nodes from a peer.

ToxUdpPacket is to be passed to Sender created by send_packets().

RequestId is to be used for tracking node timeouts.

pub fn request_nodes_close(&mut self) -> Vec<ToxUdpPacket>

Create ToxUdpPackets with request for nodes from every peer in the Close List.

ToxUdpPackets are to be passed to Sender created by send_packets().

Adds request to response timeout queue.

Note: returned Vec can be empty if there are no known nodes.

pub fn send_nodes(
&self,
sink: ToxSplitSink,
peer_addr: SocketAddr,
request: &DhtPacket
) -> Option<SendSink>

Send nodes in response to GetNodes request contained in DhtPacket.

Can fail (return None) if Kbucket is empty or DhtPacket doesn't contain GetNodes request.

pub fn handle_packet(&mut self, packet: &DhtPacket) -> Option<DhtPacket>

Function to handle incoming packets. If there is a response packet, Some(DhtPacket) is returned.