Struct tox::toxcore::network::NetworkingCore

pub struct NetworkingCore { /* fields omitted */ }

Networking Core.

Methods

impl NetworkingCore

pub fn new<R>(
ip: IpAddr,
port_range: R,
handle: &Handle
) -> Result<NetworkingCore> where
R: Into<PortRange<u16>>,

Initialize networking by binding to specified socket IP, and a port in supplied range.

Fails when

binding to IP address and every port in supplied range fails

setting broadcast on socket fails

pub fn register(
&mut self,
byte: u8,
cb: PacketHandlerCallback,
object: Rc<RefCell<Any>>
)

Function to call when packet beginning with byte is received.

pub fn poll(&self)

Call this several times a second.

pub fn send_packet(&self, addr: SocketAddr, data: &[u8]) -> Result<usize>

Function to send packet (data) to SocketAddr.

pub fn receive_packet(&self, data: &mut [u8]) -> Result<(usize, SocketAddr)>

Receive packet data into &mut data.

Returns Option<(length, addr)> or io::Error.

If successfull, (received bytes length, sender socket addr) are returned.