Module tox::toxcore::state_format::old

Old Tox State Format (TSF). Will be deprecated when something better will become available.

Structs

DhtState
DHT section of the old state format.

FriendState
Friend state format for a single friend, compatible with what C toxcore does with on GCC x86{,_x64} platform.

Friends
Wrapper struct for Vec<FriendState> to ease working with friend lists.

Name
Own name, up to NAME_LEN bytes long.

NospamKeys
NoSpam and Keys section of the old state format.

PathNodes
Contains list in PackedNode format.

State
Tox State sections. Use to manage .tox save files.

StatusMsg
Status message, up to STATUS_MSG_LEN bytes.

TcpRelays
Contains list in PackedNode format.

Enums

FriendStatus
Friend state status. Used by FriendState.

Section
Sections of state format.

SectionKind
Sections of the old state format.

UserStatus
User status. Used for both own & friend statuses.

Constants

DHT_STATE_MIN_SIZE
Minimal number of bytes DhtState has.

FRIENDSTATEBYTES
Number of bytes of serialized FriendState.

NAME_LEN
Length in bytes of name. Will be moved elsewhere.

NOSPAMKEYSBYTES
Number of bytes of serialized NospamKeys.

STATUS_MSG_LEN
Length in bytes of friend's status message.