Struct tox::toxcore::state_format::old::State

pub struct State { /* fields omitted */ }

Tox State sections. Use to manage .tox save files.

https://zetok.github.io/tox-spec/#state-format

Methods

impl State

pub fn add_friend_norequest(&mut self, pk: &PublicKey) -> bool

Add friend with PublicKey without sending a friend request.

Returns true if friend was added, false otherwise.

Subject to change.

pub fn is_own_pk(&self, pk: &PublicKey) -> bool

Check if given PublicKey is an exact match to the State PK.

When checking if given Tox ID is our own, check only PK part, as it is the only usable unchanging part.

Returns true if there's an exact match, false otherwise.