Module tox::toxcore::tcp

TCP handshake and Packet handling

Modules

binary_io
Encoding/decoding traits for TCP mod

codec
Codec implementation for encoding/decoding TCP Packets in terms of tokio-io

handshake
Handshake packets to establish a confirmed connection via handshake using Diagram

packet
Top-level TCP Packets according to Tox spec

secure
A tls-like module for relay packets

server
The implementation of TCP relay server

Functions

create_client_handshake
Create a handshake from client to server

handle_client_handshake
Handle received client handshake on the server side. Return secure::Channel, Client PK, server handshake

handle_server_handshake
Handle received server handshake on the client side.

make_client_handshake
Sends handshake to the server, receives handshake from the server and processes it

make_server_handshake
Receives handshake from the client, processes it and sends handshake to the client