Struct tox::toxcore::tcp::secure::Session

pub struct Session { /* fields omitted */ }

A Session is created on both sides. Its PK and Nonce is sent from a client to server via handshake. A server creates its own Session, creates Channel and replies with its Session PK and Nonce to the client, the client creates Channel.

After the handshake is complited, they may both encrypt and decrypt messages using their Channels.

Methods

impl Session

pub fn new() -> Session

Create a new Session with random pk, sk and nonce. You should send it to establish a secure Channel

pub fn pk(&self) -> &PublicKey

Get the PK of the Session

pub fn nonce(&self) -> &Nonce

Get the Nonce of the Session

pub fn create_precomputed_key(&self, other_pk: &PublicKey) -> PrecomputedKey

Create PrecomputedKey to encrypt/decrypt data using secure Channel