Struct tox::toxcore::tcp::server::Client

pub struct Client { /* fields omitted */ }

Structure that represents how Server keeps connected clients. A write-only socket with human interface. A client cannot send a message directly to another client, whereas server can.

Methods

impl Client

pub fn new(tx: UnboundedSender<Packet>, pk: &PublicKey) -> Client

Create new Client

pub fn pk(&self) -> PublicKey

PK of the Client

pub fn ping_id(&self) -> u64

Last ping_id sent to client.

pub fn get_connection_id(&self, to: &PublicKey) -> Option<u8>

Return index of of the link by PK

Some(index + 16) if link exists

None if there is no such PK linked to this client

pub fn insert_connection_id(&mut self, to: &PublicKey) -> Option<u8>

Try to link PK.

Some(index + 16) if has been inserted or link existed

None if no free space to insert

pub fn get_link(&self, connection_id: u8) -> Option<PublicKey>

Get link by connection_id. Ensure connection_id [0; 240) + 16

pub fn take_link(&mut self, connection_id: u8) -> Option<PublicKey>

Get link by connection_id and remove it from container. Ensure connection_id [0; 240) + 16

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter_links(&self) -> Iter<Option<PublicKey>>

Iter over each link in links of the Client

pub fn send_route_response(
&self,
pk: &PublicKey,
connection_id: u8
) -> IoFuture<()>

Construct RouteResponse and send it to Client

pub fn send_connect_notification(&self, connection_id: u8) -> IoFuture<()>

Construct ConnectNotification and send it to Client ignoring IO error

pub fn send_disconnect_notification(&self, connection_id: u8) -> IoFuture<()>

Construct DisconnectNotification and send it to Client ignoring IO error

pub fn send_pong_response(&self, ping_id: u64) -> IoFuture<()>

Construct PongResponse and send it to Client

pub fn send_oob(&self, sender_pk: &PublicKey, data: Vec<u8>) -> IoFuture<()>

Construct OobReceive and send it to Client ignoring IO error

pub fn send_data(&self, connection_id: u8, data: Vec<u8>) -> IoFuture<()>

Construct Data and send it to Client