Struct tox::toxcore::tcp::server::Server

pub struct Server { /* fields omitted */ }

A Server is a structure that holds connected clients, manages their links and handles their responses. Notice that there is no actual network code here, the Server accepts packets by value from Server::handle_packet, sends packets back to clients via futures::sync::mpsc::UnboundedSender<Packet> channel. The outer code should manage how to handshake connections, get packets from clients, pass them into Server::handle_packet, create mpsc chanel, take packets from futures::sync::mpsc::UnboundedReceiver<Packet> send them back to clients via network.

Methods

impl Server

pub fn new() -> Server

Create a new Server

pub fn insert(&self, client: Client)

Insert the client into connected_clients. Do nothing else.

pub fn handle_packet(&self, pk: &PublicKey, packet: Packet) -> IoFuture<()>

The main processing function. Call in on each incoming packet from connected and handshaked client.

pub fn shutdown_client(&self, pk: &PublicKey) -> IoFuture<()>

Gracefully shutdown client by pk. Remove it from the list of connected clients. If there are any clients mutually linked to current client, we send them corresponding DisconnectNotification.