Struct tox::toxcore::timeout::TimeoutQueue

pub struct TimeoutQueue { /* fields omitted */ }

Store & manage timeout data.

To create new TimeoutQueue use Default trait:

use tox::toxcore::timeout::TimeoutQueue;

let _timeout_queue = TimeoutQueue::default();

Methods

impl TimeoutQueue

pub fn push(&mut self, nt: NodeTimeout)

Push the timeout to queue.

pub fn add(&mut self, pk: &PublicKey, id: RequestId)

Create new timeout and add it to the queue

pub fn remove(&mut self, id: RequestId) -> bool

Remove timeout with supplied ID and all timeouts that have same PK as the removed timeout.

Returns true if there was at least 1 timeout that was removed, false otherwise.

pub fn get_timed_out(&mut self, secs: u64) -> Vec<PublicKey>

Remove from TimeoutQueue and return all PublicKeys of nodes that timed out.

Note: Returns an empty Vec if there are no nodes that have timed out.