Struct tox::toxencryptsave::PassKey

pub struct PassKey { /* fields omitted */ }

Key and Salt that are used to encrypt/decrypt data.

Methods

impl PassKey

pub fn new(passphrase: &[u8]) -> Result<PassKey, KeyDerivationError>

Create a new PassKey with a random Salt.

Note that passphrase memory is not being zeroed after it has been used. Code that provides passphrase should take care of zeroing that memory.

Can fail for the same reasons as PassKey::with_salt(), that is:

passphrase is empty

deriving key failed (can happen due to OOM)

E.g.

use self::tox::toxencryptsave::*;

// fails with an empty passphrase
assert_eq!(Err(KeyDerivationError::Null), PassKey::new(&[]));

pub fn with_salt(
passphrase: &[u8],
salt: Salt
) -> Result<PassKey, KeyDerivationError>

Create a new PassKey with provided Salt, rather than using a random one.

Note that passphrase memory is not being zeroed after it has been used. Code that provides passphrase should take care of zeroing that memory.

Fails when:

passphrase is empty

deriving key failed (can happen due to OOM)

E.g.

extern crate sodiumoxide;
extern crate tox;

use sodiumoxide::crypto::pwhash::gen_salt;
use tox::toxencryptsave::*;

assert_eq!(Err(KeyDerivationError::Null),
            PassKey::with_salt(&[], gen_salt()));

pub fn encrypt(&self, data: &[u8]) -> Result<Vec<u8>, EncryptionError>

Encrypts provided data with self PassKey.

Encrypted data is bigger than supplied data by EXTRA_LENGTH.

Fails when:

provided data is empty

E.g.

use tox::toxencryptsave::*;

                          // ↓ don't
let passkey = PassKey::new(&[0]).expect("Failed to unwrap PassKey!");

assert_eq!(Err(EncryptionError::Null), passkey.encrypt(&[]));

pub fn decrypt(&self, data: &[u8]) -> Result<Vec<u8>, DecryptionError>

Decrypts provided data with self PassKey.

Decrypted data is smaller by EXTRA_LENGTH than encrypted data.

Fails when:

provided data is empty

size of provided data is less than EXTRA_LENGTH

format of provided data is wrong

decrypting data fails

could be due to OOM or by providing bytes that aren't encrypted after encrypted part

E.g.

use self::tox::toxencryptsave::*;

                          // ↓ don't
let passkey = PassKey::new(&[0]).expect("Failed to unwrap PassKey!");

// empty data
assert_eq!(Err(DecryptionError::Null), passkey.decrypt(&[]));