Module tox::toxcore::dht

DHT part of the toxcore.

takes care of the serializing and de-serializing DHT packets

..

Structs

Bucket
Structure for holding up to BUCKET_SIZE nodes.

DhtPacket
Standard DHT packet that encapsulates in the encrypted payload DhtPacketT.

GetNodes
Request to get address of given DHT PK, or nodes that are closest in DHT to the given PK.

Kbucket
K-bucket structure to hold up to KBUCKET_MAX_ENTRIES * BUCKET_SIZE nodes close to own PK.

Node
DHT Node and its associated info.

PackedNode
Packed Node format is a way to store the node info in a small yet easy to parse format.

Ping
Used to request/respond to ping. Use in an encrypted form in DHT packets.

SendNodes
Response to GetNodes request, containing up to 4 nodes closest to the requested node.

Enums

DPacketT
Types of DHT packets that can be put in DHT Packet.

IpType
Used by PackedNode.

PacketKind
Top-level packet kind names and their associated numbers.

PingType
Type of Ping packet. Either a request or response.

Constants

BUCKET_SIZE
Maximum of nodes that Bucket can hold.

DHT_PACKET_MIN_SIZE
Minimal size of DhtPacket in bytes.

GET_NODES_SIZE
Size of serialized GetNodes in bytes.

KBUCKET_MAX_ENTRIES
Maximum number of buckets that Kbucket can hold.

PACKED_NODE_IPV4_SIZE
Size in bytes of serialized PackedNode with IPv4.

PACKED_NODE_IPV6_SIZE
Size in bytes of serialized PackedNode with IPv6.

PING_SIZE
Length in bytes of Ping when serialized into bytes.

Traits

Distance
Trait for functionality related to distance between PublicKeys.

Functions

kbucket_index
Calculate the k-bucket index of a PK compared to "own" PK.