Struct tox::toxcore::dht::Bucket

pub struct Bucket { /* fields omitted */ }

Structure for holding nodes.

Number of nodes it can contain is set during creation. If not set (aka None is supplied), number of nodes defaults to BUCKET_DEFAULT_SIZE.

Nodes stored in Bucket are in PackedNode format.

Used in Kbucket for storing nodes close to given PK; and additionally used to store nodes closest to friends.

Spec definition.

Methods

impl Bucket

pub fn new(num: Option<u8>) -> Self

Create a new Bucket to store nodes close to the pk.

Can hold up to num nodes if number is supplied. If None is supplied, holds up to BUCKET_DEFAULT_SIZE nodes. If Some(0) is supplied, it is treated as None.

pub fn try_add(&mut self, base_pk: &PublicKey, new_node: &PackedNode) -> bool

Try to add PackedNode to the bucket.

If the PackedNode with given PublicKey is already in the Bucket, the PackedNode is updated (since its SocketAddr can differ).

If bucket is not full, node is appended.

If bucket is full, node's closeness is compared to nodes already in bucket, and if it's closer than some node, it prepends that node, and last node is removed from the list.

If the node being added is farther away than the nodes in the bucket, it isn't added and false is returned.

Note that you must pass the same base_pk each call or the internal state will be undefined.

Returns true if node was added, false otherwise.

pub fn remove(&mut self, base_pk: &PublicKey, node_pk: &PublicKey)

Remove PackedNode with given PK from the Bucket.

Note that you must pass the same base_pk each call or the internal state will be undefined. Also base_pk must be equal to base_pk you added a node with. Normally you don't call this function on your own but Kbucket does.

If there's no PackedNode with given PK, nothing is being done.

pub fn capacity(&self) -> usize

Get the capacity of the Bucket.

pub fn is_empty(&self) -> bool

Check if Bucket is empty.

Returns true if there are no nodes in the Bucket, false otherwise.

pub fn is_full(&self) -> bool

Check if Bucket is full.

Returns true if there is no free space in the Bucket, false otherwise.