Struct tox::toxcore::dht::Bucket

pub struct Bucket<'a> {
    pub pk: &'a PublicKey,
    // some fields omitted
}

Structure for holding up to BUCKET_SIZE nodes.

Nodes in bucket are sorted by closeness to the PK; closes node is the first, while furthest is last.

Used in Kbucket for storing nodes close to own PK; and additionally used to store nodes closest to friends.

Fields

pk

PK that nodes in the bucket are close to.

Methods

impl<'a> Bucket<'a>

fn new(pk: &'a PublicKey, i: u8) -> Self

Create a new Bucket to store nodes close to the pk.

fn try_add(&mut self, node: &Node) -> bool

Try to add node to the bucket.

If bucket doesn't have BUCKET_SIZE nodes, node is appended.

If bucket has BUCKET_SIZE already, node's closeness is compared to nodes already in bucket, and if it's closer than some node, it prepends that node, and last node is removed from the list.

Returns true is node was added, false otherwise.