Struct tox::toxcore::dht::DhtRequest

pub struct DhtRequest {
    pub receiver: PublicKey,
    pub sender: PublicKey,
    // some fields omitted
}

DHT Request packet structure.

Used to send data via one node1 to other node2 via intermediary node when there is no direct connection between nodes 1 and 2.

<own node> → <connected, intermediary node> → <not connected node>

When receiving DhtRequest own instance should check whether receiver PK matches own PK, or PK of a known node.

if it matches own PK, handle it.

if it matches PK of a known node, send packet to that node

Serialized structure:

Length 	Contents

1 	32

32 	receiver's DHT public key

32 	sender's DHT public key

24 	Nonce

? 	encrypted data

Fields

receiver

PublicKey of receiver.

sender

PUblicKey of sender.

Methods

impl DhtRequest

fn new(secret_key: &SecretKey, own_public_key: &PublicKey, receiver_public_key: &PublicKey, nonce: &Nonce, packet: DhtRequestT) -> Self

Create a new DhtRequest.

fn get_request(&self, secret_key: &SecretKey) -> Option<DhtRequestT>

Get request data. This function decrypts payload and tries to parse it as request type.

Returns None in case of failure.