Struct tox::toxcore::dht::Kbucket

pub struct Kbucket {
    pub pk: PublicKey,
    pub k: u8,
    // some fields omitted
}

K-buckets structure to hold up to KBUCKET_MAX_ENTRIES * BUCKET_DEFAULT_SIZE nodes close to own PK.

Nodes in bucket are sorted by closeness to the PK; closest node is the first, while furthest is last.

Further reading: Tox spec.

Fields

pk

PublicKey for which Kbucket holds close nodes.

k

Number of Buckets held.

Methods

impl Kbucket

fn new(k: u8, pk: &PublicKey) -> Self

Create a new Kbucket.

k – number of Buckets held.

fn try_add(&mut self, node: &PackedNode) -> bool

Add PackedNode to Kbucket.

Node can be added only if:

its kbucket index is lower or equal to k (number of buckets).

Bucket to which it is added has free space or added node is closer to the PK than other node in the bucket.

Returns true if node was added successfully, false otherwise.

fn remove(&mut self, pk: &PublicKey)

Remove PackedNode with given PK from the Kbucket.

fn get_closest(&self, pk: &PublicKey) -> Vec<PackedNode>

Get (up to) 4 closest nodes to given PK.

Functionality for SendNodes.

Returns less than 4 nodes only if Kbucket contains less than 4 nodes.