Struct tox::toxcore::dht::PackedNode

pub struct PackedNode {
    pub ip_type: IpType,
    pub saddr: SocketAddr,
    pub pk: PublicKey,
}

Packed Node format is a way to store the node info in a small yet easy to parse format.

It is used in many places in Tox, e.g. in DHT Send nodes.

To store more than one node, simply append another on to the previous one:

[packed node 1][packed node 2][...]

Packed node format:

                         (39 bytes for IPv4, 51 for IPv6)
+-----------------------------------+
| ip_type                ( 1 byte ) |
|                                   |
| IPv4 Address           ( 4 bytes) |
|  -OR-                     -OR-    |
| IPv6 Address           (16 bytes) |
| Port                   ( 2 bytes) |
| Node ID                (32 bytes) |
+-----------------------------------+

DHT module should use only UDP variants of IpType, given that DHT runs solely on the UDP.

Fields

ip_type

IP type, includes also info about protocol used.

saddr

Socket addr of node.

pk

Public Key of the node.

Methods

impl PackedNode

fn new(udp: bool, saddr: SocketAddr, pk: &PublicKey) -> Self

New PackedNode.

udp – whether UDP or TCP should be used. UDP is used for DHT nodes, whereas TCP is used for TCP relays.

fn ip(&self) -> IpAddr

Get an IP address from the PackedNode.

fn from_bytes_multiple(bytes: &[u8]) -> Option<Vec<PackedNode>>

Parse bytes into multiple PackedNodes.

If provided bytes are smaller than PACKED_NODE_IPV4_SIZE or can't be parsed, returns None.

Parses nodes until first error is encountered.