Struct tox::toxcore::network::NetworkingCore

pub struct NetworkingCore { /* fields omitted */ }

Networking Core.

Methods

impl NetworkingCore

fn new<R: Into<PortRange<u16>>>(ip: IpAddr,
port_range: R)
-> Result<NetworkingCore>

Initialize networking by binding to specified socket IP, and a port in supplied range.

Fails when

binding to IP address and every port in supplied range fails

setting broadcast on socket fails

use tox::toxcore::network::NetworkingCore;

NetworkingCore::new("::".parse().unwrap(), 33445..33545).unwrap();

fn register(&mut self,
byte: u8,
cb: PacketHandlerCallback,
object: Rc<RefCell<Any>>)

Function to call when packet beginning with byte is received.

   fn callback(num: Rc<RefCell<Any>>, _: SocketAddr, _: &[u8]) -> usize {
       match num.borrow().downcast_ref::<usize>() {
           Some(&num) => unimplemented!(),
           None => 0
       }
   }

   net.register(99, callback, Rc::new(RefCell::new(1usize)) as Rc<RefCell<Any>>);

fn poll(&self)

Call this several times a second.

fn send_packet(&self, addr: SocketAddr, data: &[u8]) -> Result<Option<usize>>

Function to send packet (data) to SocketAddr.

fn receive_packet(&self, data: &mut [u8]) -> Result<Option<(usize, SocketAddr)>>

Receive packet data into &mut data.

Returns Option<(length, addr)> or io::Error.

If successfull, (received bytes length, sender socket addr) are returned.