Struct tox::toxcore::toxid::NoSpam

pub struct NoSpam(pub [u8; NOSPAMBYTES]);

NoSpam used in ToxId.

Number is used to make sure that there is no friend requests from peers that know out long term PK, but don't actually know Tox ID.

The preferred way of creating NoSpam is to generate a random one.

Additionally, it should be possible to set a custom NoSpam.

https://zetok.github.io/tox-spec/#tox-id

Methods

impl NoSpam

fn new() -> Self

Create new NoSpam with random bytes.

Two new() NoSpams will always be different:

use self::tox::toxcore::toxid::NoSpam;

assert!(NoSpam::new() != NoSpam::new());